requirements.txt
# Core Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0

# LLM - Groq (free tier)
groq>=0.4.0,<1.0.0

# Web Search - DuckDuckGo (free, no API key)
duckduckgo-search>=4.0.0,<8.0.0

# Web Scraping (free)
httpx>=0.26.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0

# Agent Orchestration
langgraph>=0.0.30

# Environment
python-dotenv>=1.0.0,<2.0.0